some refactoring of variables
